Analysis/requirements.txt
<<<<<<< HEAD
hid>=1.0.0
pyusb>=1.2.0
pyserial>=3.5


=======
hid>=1.0.0
pyusb>=1.2.0
pyserial>=3.5


>>>>>>> 98d193724cd3ec1bfca0d21e059c3a30b2c9dd50
